Improved naming conventions and created first client-server communication.
