--- FrontEnd/src/pages/QuizMaker.tsx
@@ -1,5 +1,0 @@
-const QuizMaker = () => {
-  return <>Hello Quiz Maker!</>;
-};
-
-export default QuizMaker;
